feat: Regrouper lles types pour permettre l'héritage et l'imbrication de "AND" dans les "OU" refers to https://trello.com/c/3zqUkzR1/2480
