+++ packages/api/server/models/_common/types/Where.ts
@@ -1,3 +1,17 @@
-import { WhereClauseGroup } from './WhereClauseGroup';
+import { Primitive } from './Primitive';
+
+export type WhereClause = {
+    value?: Primitive | Primitive[],
+    query?: string,
+    operator?: string,
+    anyOperator?: string,
+    arrayOperator?: boolean,
+    not?: boolean,
+    andClauses?: Where
+};
+
+export type WhereClauseGroup = {
+    [column: string]: WhereClause
+};
 
 export type Where = Array<WhereClauseGroup>;
