--- packages/api/server/models/_common/types/WhereClause.ts
@@ -1,11 +1,0 @@
-
-import { Primitive } from './Primitive';
-
-export type WhereClause = {
-    value: Primitive | Primitive[],
-    query?: string,
-    operator?: string,
-    anyOperator?: string,
-    arrayOperator?: boolean,
-    not?: boolean
-};
