--- packages/api/server/models/_common/types/WhereClauseGroup.ts
@@ -1,6 +1,0 @@
-
-import { WhereClause } from './WhereClause';
-
-export type WhereClauseGroup = {
-    [column: string]: WhereClause
-};
