+++ packages/api/server/utils/permission/where.ts
@@ -1,4 +1,4 @@
-import { WhereClauseGroup } from '#server/models/_common/types/WhereClauseGroup';
+import { WhereClauseGroup } from '#server/models/_common/types/Where';
 import getPermission from './getPermission';
 import { User } from '#root/types/resources/User.d';
